refactor(client): create new component to ApplyChanges dialog BTM
