+++ packages/amplication-client/src/Project/ArchitectureConsole/ModelOrganizer.tsx
@@ -38,6 +38,7 @@ import {
   Node,
   NodePayloadWithPayloadType,
 } from "./types";
+import ApplyChangesConfirmationModelOrganizer from "./ApplyChangesConfirmationModelOrganizer";
 
 export const CLASS_NAME = "model-organizer";
 const REACT_FLOW_CLASS_NAME = "reactflow-wrapper";
@@ -98,12 +99,6 @@ export default function ModelOrganizer() {
   const [currentDropTarget, setCurrentDropTarget] = useState<Node>(null);
 
   const [isValidResourceName, setIsValidResourceName] = useState<boolean>(true);
-  const [
-    saveConfirmApplyChangesTampTrigger,
-    setSaveConfirmApplyChangesTampTrigger,
-  ] = useState<Date>(null);
-  const [confirmApplyChangesDialog, setConfirmApplyChangesDialog] =
-    useState<boolean>(false);
 
   const fitViewTimerRef = useRef(null);
 
@@ -127,15 +122,6 @@ export default function ModelOrganizer() {
     return () => clearTimeout(fitViewTimerRef.current);
   }, []);
 
-  useEffect(() => {
-    if (
-      redesignMode &&
-      !saveConfirmApplyChangesTampTrigger &&
-      (changes?.movedEntities?.length > 0 || changes?.newServices?.length > 0)
-    )
-      setConfirmApplyChangesDialog(true);
-  }, [redesignMode, saveConfirmApplyChangesTampTrigger, changes]);
-
   const onNodesChange = useCallback(
     (changes) => {
       const updatedNodes = applyNodeChanges<NodePayloadWithPayloadType>(
@@ -146,11 +132,6 @@ export default function ModelOrganizer() {
     },
     [nodes, setNodes]
   );
-
-  const onApplyChangesConfirmationClicked = useCallback(() => {
-    setConfirmApplyChangesDialog(false);
-    setSaveConfirmApplyChangesTampTrigger(new Date());
-  }, [setConfirmApplyChangesDialog]);
 
   const handleCreateResourceState = useCallback(() => {
     setIsValidResourceName(true);
@@ -304,15 +285,9 @@ export default function ModelOrganizer() {
             />
           </div>
           <div className={`${CLASS_NAME}__body`}>
-            <ConfirmationDialog
-              isOpen={confirmApplyChangesDialog}
-              onDismiss={onApplyChangesConfirmationClicked}
-              message={`Your architecture tweaks are ready to be applied, allowing Amplication to generate its code.
-              To reset or fetch updates not in the current state, use the buttons in the top toolbar.`}
-              confirmButton={{ label: "Got it" }}
-              onConfirm={onApplyChangesConfirmationClicked}
-            ></ConfirmationDialog>
-
+            <ApplyChangesConfirmationModelOrganizer
+              changes={changes}
+            ></ApplyChangesConfirmationModelOrganizer>
             <ModelOrganizerToolbar
               changes={changes}
               nodes={nodes}
